pyproject.toml
[tool.poetry]
name = "jobby-bot"
version = "0.1.0"
description = "AI-powered job application automation using Claude Agent SDK"
authors = ["Anmol Desai <adcan288@gmail.com>"]
readme = "README.md"
packages = [{include = "jobby_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
claude-agent-sdk = "^0.1.0"
python-jobspy = "*"
notion-client = "^2.2.1"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
pandas = "^2.0.0"
weasyprint = "^60.0"
pyyaml = "^6.0"
markdown = "^3.5"
jinja2 = "^3.1.0"
anthropic = "^0.39.0"
pdfplumber = "^0.11.0"
reportlab = "^4.0.0"
"discord.py" = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.12.0"
isort = "^5.13.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
